feat: Add logic to manage a virtual default profile
This profile will not be saved in the profiles file but will show up when
profiles are listed and will allow users to switch to the default profile.

We still need to figure out how to disable this if there is a local profile
from the desktop app.

https://github.com/contextforge-org/contextforge-cli/issues/13
Branch: Profiles-13

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>